fix; 암묵적 형변환을 지양하기 위한 명시적 형변환 추가

+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/common/security/authorization/SpendingManager.java
@@ -26,7 +26,7 @@ public class SpendingManager {
 
     @Transactional(readOnly = true)
     public boolean hasPermissions(Long userId, List<Long> spendingIds) {
-        if (spendingService.countByUserIdAndIdIn(userId, spendingIds) != spendingIds.size()) {
+        if (spendingService.countByUserIdAndIdIn(userId, spendingIds) != (long) spendingIds.size()) {
             return false;
         }
